@@ -17,12 +17,13 @@
         }
     ],
     "require": {
-        "php": "^7.1",
-        "psr/log": "^1.0",
-        "symfony/var-dumper": "^2.6|^3|^4"
+        "php": "^7.1|^8",
+        "psr/log": "^1|^2|^3",
+        "symfony/var-dumper": "^4|^5|^6|^7"
     },
     "require-dev": {
-        "phpunit/phpunit": "^5"
+        "phpunit/phpunit": ">=7.5.20 <10.0",
+        "twig/twig": "^1.38|^2.7|^3.0"
     },
     "autoload": {
         "psr-4": {
@@ -36,7 +37,7 @@
     },
      "extra": {
         "branch-alias": {
-            "dev-master": "1.16-dev"
+            "dev-master": "1.20-dev"
         }
     }
 }
