@@ -10,68 +10,46 @@
 
 namespace DebugBar\DataCollector;
 
-use DebugBar\DataFormatter\DataFormatter;
-use DebugBar\DataFormatter\DataFormatterInterface;
-use DebugBar\DataFormatter\DebugBarVarDumper;
+use DebugBar\DataFormatter\HasDataFormatter;
 
 /**
  * Abstract class for data collectors
  */
 abstract class DataCollector implements DataCollectorInterface
 {
-    private static $defaultDataFormatter;
-    private static $defaultVarDumper;
+    use HasDataFormatter;
 
-    protected $dataFormater;
-    protected $varDumper;
+    public static $defaultDataFormatter;
+    public static $defaultVarDumper;
+
     protected $xdebugLinkTemplate = '';
     protected $xdebugShouldUseAjax = false;
     protected $xdebugReplacements = array();
 
     /**
-     * Sets the default data formater instance used by all collectors subclassing this class
+     * Shorten the file path by removing the xdebug path replacements
      *
-     * @param DataFormatterInterface $formater
+     * @param string $file
+     * @return string
      */
-    public static function setDefaultDataFormatter(DataFormatterInterface $formater)
+    public function normalizeFilePath($file)
     {
-        self::$defaultDataFormatter = $formater;
-    }
-
-    /**
-     * Returns the default data formater
-     *
-     * @return DataFormatterInterface
-     */
-    public static function getDefaultDataFormatter()
-    {
-        if (self::$defaultDataFormatter === null) {
-            self::$defaultDataFormatter = new DataFormatter();
+        if (empty($file)) {
+            return '';
         }
-        return self::$defaultDataFormatter;
-    }
 
-    /**
-     * Sets the data formater instance used by this collector
-     *
-     * @param DataFormatterInterface $formater
-     * @return $this
-     */
-    public function setDataFormatter(DataFormatterInterface $formater)
-    {
-        $this->dataFormater = $formater;
-        return $this;
-    }
-
-    /**
-     * @return DataFormatterInterface
-     */
-    public function getDataFormatter()
-    {
-        if ($this->dataFormater === null) {
-            $this->dataFormater = self::getDefaultDataFormatter();
+        if (@file_exists($file)) {
+            $file = realpath($file);
         }
-        return $this->dataFormater;
+
+        foreach (array_keys($this->xdebugReplacements) as $path) {
+            if (strpos($file, $path) === 0) {
+            	$file = substr($file, strlen($path));
+                break;
+            }
+        }
+
+        return ltrim(str_replace('\\', '/', $file), '/');
     }
 
     /**
@@ -87,87 +65,28 @@ abstract class DataCollector implements DataCollectorInterface
      */
     public function getXdebugLink($file, $line = 1)
     {
-        if (count($this->xdebugReplacements)) {
-            $file = strtr($file, $this->xdebugReplacements);
+        if (empty($file)) {
+            return null;
         }
 
-        $url = strtr($this->getXdebugLinkTemplate(), ['%f' => $file, '%l' => $line]);
+        if (@file_exists($file)) {
+            $file = realpath($file);
+        }
+
+        foreach ($this->xdebugReplacements as $path => $replacement) {
+            if (strpos($file, $path) === 0) {
+            	$file = $replacement . substr($file, strlen($path));
+            	break;
+            }
+        }
+
+        $url = strtr($this->getXdebugLinkTemplate(), [
+            '%f' => rawurlencode(str_replace('\\', '/', $file)),
+            '%l' => rawurlencode((string) $line),
+        ]);
         if ($url) {
             return ['url' => $url, 'ajax' => $this->getXdebugShouldUseAjax()];
         }
-    }
-  
-    /**  
-     * Sets the default variable dumper used by all collectors subclassing this class
-     *
-     * @param DebugBarVarDumper $varDumper
-     */
-    public static function setDefaultVarDumper(DebugBarVarDumper $varDumper)
-    {
-        self::$defaultVarDumper = $varDumper;
-    }
-
-    /**
-     * Returns the default variable dumper
-     *
-     * @return DebugBarVarDumper
-     */
-    public static function getDefaultVarDumper()
-    {
-        if (self::$defaultVarDumper === null) {
-            self::$defaultVarDumper = new DebugBarVarDumper();
-        }
-        return self::$defaultVarDumper;
-    }
-
-    /**
-     * Sets the variable dumper instance used by this collector
-     *
-     * @param DebugBarVarDumper $varDumper
-     * @return $this
-     */
-    public function setVarDumper(DebugBarVarDumper $varDumper)
-    {
-        $this->varDumper = $varDumper;
-        return $this;
-    }
-
-    /**
-     * Gets the variable dumper instance used by this collector; note that collectors using this
-     * instance need to be sure to return the static assets provided by the variable dumper.
-     *
-     * @return DebugBarVarDumper
-     */
-    public function getVarDumper()
-    {
-        if ($this->varDumper === null) {
-            $this->varDumper = self::getDefaultVarDumper();
-        }
-        return $this->varDumper;
-    }
-
-    /**
-     * @deprecated
-     */
-    public function formatVar($var)
-    {
-        return $this->getDataFormatter()->formatVar($var);
-    }
-
-    /**
-     * @deprecated
-     */
-    public function formatDuration($seconds)
-    {
-        return $this->getDataFormatter()->formatDuration($seconds);
-    }
-
-    /**
-     * @deprecated
-     */
-    public function formatBytes($size, $precision = 2)
-    {
-        return $this->getDataFormatter()->formatBytes($size, $precision);
     }
 
     /**
@@ -180,6 +99,39 @@ abstract class DataCollector implements DataCollectorInterface
         }
 
         return $this->xdebugLinkTemplate;
+    }
+
+    /**
+     * @param string $editor
+     */
+    public function setEditorLinkTemplate($editor)
+    {
+        $editorLinkTemplates = array(
+            'sublime' => 'subl://open?url=file://%f&line=%l',
+            'textmate' => 'txmt://open?url=file://%f&line=%l',
+            'emacs' => 'emacs://open?url=file://%f&line=%l',
+            'macvim' => 'mvim://open/?url=file://%f&line=%l',
+            'phpstorm' => 'phpstorm://open?file=%f&line=%l',
+            'phpstorm-remote' => 'javascript:(()=>{let r=new XMLHttpRequest;' .
+                'r.open(\'get\',\'http://localhost:63342/api/file/%f:%l\');r.send();})()',
+            'idea' => 'idea://open?file=%f&line=%l',
+            'idea-remote' => 'javascript:(()=>{let r=new XMLHttpRequest;' .
+                'r.open(\'get\',\'http://localhost:63342/api/file/?file=%f&line=%l\');r.send();})()',
+            'vscode' => 'vscode://file/%f:%l',
+            'vscode-insiders' => 'vscode-insiders://file/%f:%l',
+            'vscode-remote' => 'vscode://vscode-remote/%f:%l',
+            'vscode-insiders-remote' => 'vscode-insiders://vscode-remote/%f:%l',
+            'vscodium' => 'vscodium://file/%f:%l',
+            'nova' => 'nova://core/open/file?filename=%f&line=%l',
+            'xdebug' => 'xdebug://%f@%l',
+            'atom' => 'atom://core/open/file?filename=%f&line=%l',
+            'espresso' => 'x-espresso://open?filepath=%f&lines=%l',
+            'netbeans' => 'netbeans://open/?f=%f:%l',
+        );
+
+        if (is_string($editor) && isset($editorLinkTemplates[$editor])) {
+            $this->setXdebugLinkTemplate($editorLinkTemplates[$editor]);
+        }
     }
 
     /**
@@ -222,11 +174,25 @@ abstract class DataCollector implements DataCollectorInterface
     /**
      * @param array $xdebugReplacements
      */
+    public function addXdebugReplacements($xdebugReplacements)
+    {
+        foreach ($xdebugReplacements as $serverPath => $replacement) {
+            $this->setXdebugReplacement($serverPath, $replacement);
+        }
+    }
+
+    /**
+     * @param array $xdebugReplacements
+     */
     public function setXdebugReplacements($xdebugReplacements)
     {
         $this->xdebugReplacements = $xdebugReplacements;
     }
 
+    /**
+     * @param string $serverPath
+     * @param string $replacement
+     */
     public function setXdebugReplacement($serverPath, $replacement)
     {
         $this->xdebugReplacements[$serverPath] = $replacement;
