@@ -20,11 +20,13 @@ class PDOCollector extends DataCollector implements Renderable, AssetProvider
 
     protected $sqlQuotationChar = '<>';
 
+    protected $durationBackground = false;
+
     /**
-     * @param TraceablePDO $pdo
+     * @param \PDO $pdo
      * @param TimeDataCollector $timeCollector
      */
-    public function __construct(TraceablePDO $pdo = null, TimeDataCollector $timeCollector = null)
+    public function __construct(\PDO $pdo = null, TimeDataCollector $timeCollector = null)
     {
         $this->timeCollector = $timeCollector;
         if ($pdo !== null) {
@@ -41,6 +43,16 @@ class PDOCollector extends DataCollector implements Renderable, AssetProvider
     {
         $this->renderSqlWithParams = $enabled;
         $this->sqlQuotationChar = $quotationChar;
+    }
+
+    /**
+     * Enable/disable the shaded duration background on queries
+     *
+     * @param  bool $enabled
+     */
+    public function setDurationBackground($enabled = true)
+    {
+        $this->durationBackground = $enabled;
     }
 
     /**
@@ -65,10 +77,13 @@ class PDOCollector extends DataCollector implements Renderable, AssetProvider
      * @param TraceablePDO $pdo
      * @param string $name Optional connection name
      */
-    public function addConnection(TraceablePDO $pdo, $name = null)
+    public function addConnection(\PDO $pdo, $name = null)
     {
         if ($name === null) {
             $name = spl_object_hash($pdo);
+        }
+        if (!($pdo instanceof TraceablePDO)) {
+            $pdo = new TraceablePDO($pdo);
         }
         $this->connections[$name] = $pdo;
     }
@@ -153,11 +168,29 @@ class PDOCollector extends DataCollector implements Renderable, AssetProvider
             }
         }
 
+        $totalTime = $pdo->getAccumulatedStatementsDuration();
+        if ($this->durationBackground && $totalTime > 0) {
+            // For showing background measure on Queries tab
+            $start_percent = 0;
+            foreach ($stmts as $i => $stmt) {
+                if (!isset($stmt['duration'])) {
+                    continue;
+                }
+
+                $width_percent = $stmt['duration'] / $totalTime * 100;
+                $stmts[$i] = array_merge($stmt, [
+                    'start_percent' => round($start_percent, 3),
+                    'width_percent' => round($width_percent, 3),
+                ]);
+                $start_percent += $width_percent;
+            }
+        }
+
         return array(
             'nb_statements' => count($stmts),
             'nb_failed_statements' => count($pdo->getFailedExecutedStatements()),
-            'accumulated_duration' => $pdo->getAccumulatedStatementsDuration(),
-            'accumulated_duration_str' => $this->getDataFormatter()->formatDuration($pdo->getAccumulatedStatementsDuration()),
+            'accumulated_duration' => $totalTime,
+            'accumulated_duration_str' => $this->getDataFormatter()->formatDuration($totalTime),
             'memory_usage' => $pdo->getMemoryUsage(),
             'memory_usage_str' => $this->getDataFormatter()->formatBytes($pdo->getPeakMemoryUsage()),
             'peak_memory_usage' => $pdo->getPeakMemoryUsage(),
@@ -171,7 +204,7 @@ class PDOCollector extends DataCollector implements Renderable, AssetProvider
      */
     public function getName()
     {
-        return 'pdo';
+        return 'database';
     }
 
     /**
@@ -179,15 +212,17 @@ class PDOCollector extends DataCollector implements Renderable, AssetProvider
      */
     public function getWidgets()
     {
+        $name = $this->getName();
+
         return array(
-            "database" => array(
+            "$name" => array(
                 "icon" => "database",
                 "widget" => "PhpDebugBar.Widgets.SQLQueriesWidget",
-                "map" => "pdo",
+                "map" => "$name",
                 "default" => "[]"
             ),
-            "database:badge" => array(
-                "map" => "pdo.nb_statements",
+            "$name:badge" => array(
+                "map" => "$name.nb_statements",
                 "default" => 0
             )
         );
