@@ -39,11 +39,12 @@ class TraceablePDOStatement extends PDOStatement
      * @param  mixed $driverdata [optional] Optional parameter(s) for the driver.
      * @return bool  TRUE on success or FALSE on failure.
      */
+    #[\ReturnTypeWillChange]
     public function bindColumn($column, &$param, $type = null, $maxlen = null, $driverdata = null)
     {
         $this->boundParameters[$column] = $param;
         $args = array_merge([$column, &$param], array_slice(func_get_args(), 2));
-        return call_user_func_array(['parent', 'bindColumn'], $args);
+        return parent::bindColumn(...$args);
     }
 
     /**
@@ -61,11 +62,11 @@ class TraceablePDOStatement extends PDOStatement
      * @param  mixed $driver_options [optional]
      * @return bool TRUE on success or FALSE on failure.
      */
-    public function bindParam($parameter, &$variable, $data_type = PDO::PARAM_STR, $length = null, $driver_options = null)
+    public function bindParam($parameter, &$variable, $data_type = PDO::PARAM_STR, $length = null, $driver_options = null) : bool
     {
         $this->boundParameters[$parameter] = $variable;
         $args = array_merge([$parameter, &$variable], array_slice(func_get_args(), 2));
-        return call_user_func_array(['parent', 'bindParam'], $args);
+        return parent::bindParam(...$args);
     }
 
     /**
@@ -80,10 +81,10 @@ class TraceablePDOStatement extends PDOStatement
      * constants.
      * @return bool TRUE on success or FALSE on failure.
      */
-    public function bindValue($parameter, $value, $data_type = PDO::PARAM_STR)
+    public function bindValue($parameter, $value, $data_type = PDO::PARAM_STR) : bool
     {
         $this->boundParameters[$parameter] = $value;
-        return call_user_func_array(['parent', 'bindValue'], func_get_args());
+        return parent::bindValue(...func_get_args());
     }
 
     /**
@@ -96,7 +97,7 @@ class TraceablePDOStatement extends PDOStatement
      * @throws PDOException
      * @return bool TRUE on success or FALSE on failure.
      */
-    public function execute($input_parameters = null)
+    public function execute($input_parameters = null) : bool
     {
         $preparedId = spl_object_hash($this);
         $boundParameters = $this->boundParameters;
