@@ -27,12 +27,14 @@ class TracedStatement
 
     protected $exception;
 
+    protected $preparedId;
+
     /**
      * @param string $sql
      * @param array $params
-     * @param string $preparedId
+     * @param null|string $preparedId
      */
-    public function __construct($sql, array $params = [], $preparedId = null)
+    public function __construct(string $sql, array $params = [], ?string $preparedId = null)
     {
         $this->sql = $sql;
         $this->parameters = $this->checkParameters($params);
@@ -43,7 +45,7 @@ class TracedStatement
      * @param null $startTime
      * @param null $startMemory
      */
-    public function start($startTime = null, $startMemory = null)
+    public function start($startTime = null, $startMemory = null) : void
     {
         $this->startTime = $startTime ?: microtime(true);
         $this->startMemory = $startMemory ?: memory_get_usage(false);
@@ -55,7 +57,7 @@ class TracedStatement
      * @param float $endTime
      * @param int $endMemory
      */
-    public function end(\Exception $exception = null, $rowCount = 0, $endTime = null, $endMemory = null)
+    public function end(\Exception $exception = null, int $rowCount = 0, float $endTime = null, int $endMemory = null) : void
     {
         $this->endTime = $endTime ?: microtime(true);
         $this->duration = $this->endTime - $this->startTime;
@@ -71,10 +73,10 @@ class TracedStatement
      * @param array $params
      * @return array
      */
-    public function checkParameters($params)
+    public function checkParameters(array $params) : array
     {
         foreach ($params as &$param) {
-            if (!mb_check_encoding($param, 'UTF-8')) {
+            if (!mb_check_encoding($param ?? '', 'UTF-8')) {
                 $param = '[BINARY DATA]';
             }
         }
@@ -86,7 +88,7 @@ class TracedStatement
      *
      * @return string
      */
-    public function getSql()
+    public function getSql() : string
     {
         return $this->sql;
     }
@@ -97,7 +99,7 @@ class TracedStatement
      * @param string $quotationChar
      * @return string
      */
-    public function getSqlWithParams($quotationChar = '<>')
+    public function getSqlWithParams(string $quotationChar = '<>') : string
     {
         if (($l = strlen($quotationChar)) > 1) {
             $quoteLeft = substr($quotationChar, 0, $l / 2);
@@ -143,7 +145,7 @@ class TracedStatement
      *
      * @return int
      */
-    public function getRowCount()
+    public function getRowCount() : int
     {
         return $this->rowCount;
     }
@@ -153,11 +155,11 @@ class TracedStatement
      *
      * @return array
      */
-    public function getParameters()
+    public function getParameters() : array
     {
         $params = [];
         foreach ($this->parameters as $name => $param) {
-            $params[$name] = htmlentities($param, ENT_QUOTES, 'UTF-8', false);
+            $params[$name] = htmlentities($param?:"", ENT_QUOTES, 'UTF-8', false);
         }
         return $params;
     }
@@ -165,9 +167,9 @@ class TracedStatement
     /**
      * Returns the prepared statement id
      *
-     * @return string
+     * @return null|string
      */
-    public function getPreparedId()
+    public function getPreparedId() : ?string
     {
         return $this->preparedId;
     }
@@ -177,7 +179,7 @@ class TracedStatement
      *
      * @return boolean
      */
-    public function isPrepared()
+    public function isPrepared() : bool
     {
         return $this->preparedId !== null;
     }
@@ -185,7 +187,7 @@ class TracedStatement
     /**
      * @return float
      */
-    public function getStartTime()
+    public function getStartTime() : float
     {
         return $this->startTime;
     }
@@ -193,7 +195,7 @@ class TracedStatement
     /**
      * @return float
      */
-    public function getEndTime()
+    public function getEndTime() : float
     {
         return $this->endTime;
     }
@@ -203,7 +205,7 @@ class TracedStatement
      *
      * @return float
      */
-    public function getDuration()
+    public function getDuration() : float
     {
         return $this->duration;
     }
@@ -211,7 +213,7 @@ class TracedStatement
     /**
      * @return int
      */
-    public function getStartMemory()
+    public function getStartMemory() : int
     {
         return $this->startMemory;
     }
@@ -219,7 +221,7 @@ class TracedStatement
     /**
      * @return int
      */
-    public function getEndMemory()
+    public function getEndMemory() : int
     {
         return $this->endMemory;
     }
@@ -229,7 +231,7 @@ class TracedStatement
      *
      * @return int
      */
-    public function getMemoryUsage()
+    public function getMemoryUsage() : int
     {
         return $this->memoryDelta;
     }
@@ -239,7 +241,7 @@ class TracedStatement
      *
      * @return boolean
      */
-    public function isSuccess()
+    public function isSuccess() : bool
     {
         return $this->exception === null;
     }
@@ -249,8 +251,8 @@ class TracedStatement
      *
      * @return \Exception
      */
-    public function getException()
-    {
+    public function getException() : \Exception
+	{
         return $this->exception;
     }
 
@@ -269,7 +271,7 @@ class TracedStatement
      *
      * @return string
      */
-    public function getErrorMessage()
+    public function getErrorMessage() : string
     {
         return $this->exception !== null ? $this->exception->getMessage() : '';
     }
