@@ -15,6 +15,10 @@ use Symfony\Component\VarDumper\Dumper\CliDumper;
 
 class DataFormatter implements DataFormatterInterface
 {
+    public $cloner;
+
+    public $dumper;
+
     /**
      * DataFormatter constructor.
      */
@@ -54,8 +58,10 @@ class DataFormatter implements DataFormatterInterface
     {
         if ($seconds < 0.001) {
             return round($seconds * 1000000) . 'μs';
-        } elseif ($seconds < 1) {
+        } elseif ($seconds < 0.1) {
             return round($seconds * 1000, 2) . 'ms';
+        } elseif ($seconds < 1) {
+            return round($seconds * 1000) . 'ms';
         }
         return round($seconds, 2) . 's';
     }
@@ -76,6 +82,25 @@ class DataFormatter implements DataFormatterInterface
 
         $base = log($size) / log(1024);
         $suffixes = array('B', 'KB', 'MB', 'GB', 'TB');
-        return $sign . round(pow(1024, $base - floor($base)), $precision) . $suffixes[floor($base)];
+        return $sign . round(pow(1024, $base - floor($base)), $precision) . $suffixes[(int) floor($base)];
+    }
+
+    /**
+     * @param object $object
+     * @return string
+     */
+    public function formatClassName($object)
+    {
+        $class = \get_class($object);
+
+        if (false === ($pos = \strpos($class, "@anonymous\0"))) {
+            return $class;
+        }
+
+        if (false === ($parent = \get_parent_class($class))) {
+            return \substr($class, 0, $pos + 10);
+        }
+
+        return $parent . '@anonymous';
     }
 }
