@@ -7,7 +7,10 @@ use DebugBar\DataCollector\TimeDataCollector;
 
 class TimeDataCollectorTest extends DebugBarTestCase
 {
-    public function setUp()
+    private $s;
+    private $c;
+
+    public function setUp(): void
     {
         $this->s = microtime(true);
         $this->c = new TimeDataCollector($this->s);
@@ -27,6 +30,7 @@ class TimeDataCollectorTest extends DebugBarTestCase
     public function testStartStopMeasure()
     {
         $this->c->startMeasure('foo', 'bar', 'baz');
+        usleep(1000);
         $this->c->stopMeasure('foo', array('bar' => 'baz'));
         $m = $this->c->getMeasures();
         $this->assertCount(1, $m);
@@ -44,5 +48,18 @@ class TimeDataCollectorTest extends DebugBarTestCase
         $this->assertTrue($data['end'] > $this->s);
         $this->assertTrue($data['duration'] > 0);
         $this->assertCount(2, $data['measures']);
+    }
+
+    public function testMeasure()
+    {
+        $returned = $this->c->measure('bar', function() {
+            usleep(50);
+            return 'returnedValue';
+        });
+        $m = $this->c->getMeasures();
+        $this->assertCount(1, $m);
+        $this->assertEquals('bar', $m[0]['label']);
+        $this->assertTrue($m[0]['start'] < $m[0]['end']);
+        $this->assertSame('returnedValue', $returned);
     }
 }
