@@ -7,14 +7,35 @@ use DebugBar\Storage\FileStorage;
 
 class FileStorageTest extends DebugBarTestCase
 {
-    public function setUp()
+    private $dirname;
+    private $s;
+    private $data;
+
+    public function setUp(): void
     {
-        $this->dirname = '/tmp/debugbar';
-        if (!file_exists($this->dirname)) {
-            mkdir($this->dirname, 0777);
+        $this->dirname = tempnam(sys_get_temp_dir(), 'debugbar');
+        if (file_exists($this->dirname)) {
+          unlink($this->dirname);
         }
+        mkdir($this->dirname, 0777);
         $this->s = new FileStorage($this->dirname);
         $this->data = array('__meta' => array('id' => 'foo'));
+        $this->s->save('bar', $this->data);
+    }
+
+    public function teardown(): void
+    {
+        $files = new \RecursiveIteratorIterator(
+            new \RecursiveDirectoryIterator($this->dirname, \RecursiveDirectoryIterator::SKIP_DOTS),
+            \RecursiveIteratorIterator::CHILD_FIRST
+        );
+
+        foreach ($files as $fileinfo) {
+            $todo = ($fileinfo->isDir() ? 'rmdir' : 'unlink');
+            $todo($fileinfo->getRealPath());
+		}
+
+        rmdir($this->dirname);
     }
 
     public function testSave()
@@ -26,7 +47,7 @@ class FileStorageTest extends DebugBarTestCase
 
     public function testGet()
     {
-        $data = $this->s->get('foo');
+        $data = $this->s->get('bar');
         $this->assertEquals($this->data, $data);
     }
 
@@ -39,6 +60,12 @@ class FileStorageTest extends DebugBarTestCase
     public function testClear()
     {
         $this->s->clear();
-        $this->assertFileNotExists($this->dirname . '/foo.json');
+
+        // avoid depreciation message on newer PHPUnit versions.  Can be removed after
+        if (method_exists($this, 'assertFileDoesNotExist')) {
+          $this->assertFileDoesNotExist($this->dirname . '/foo.json');
+        } else {
+          $this->assertFileNotExists($this->dirname . '/foo.json');
+        }
     }
 }
